py mbp: Split test into `plant_test` and `math_test`
